requirements.txt
aeon==0.5.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
confection==0.1.3
cymem==2.0.8
deprecated==1.2.14
docker-pycreds==0.4.0
fastai==2.7.13
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.19.0
fastprogress==1.0.3
fonttools==4.44.0
frozenlist==1.4.0
future==0.18.3
gitdb==4.0.11
gitpython==3.1.40
gpboost==1.2.7.1
heartpy==1.2.7
hmmlearn==0.3.0
hyperopt==0.2.7
idna==3.4
imbalanced-learn==0.11.0
ipywidgets==8.1.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyterlab-widgets==3.0.9
langcodes==3.3.0
#lightgbm==4.1.0
lightgbm
lightning==2.1.0
lightning-utilities==0.9.0
llvmlite==0.41.1
markupsafe==2.1.2
matplotlib==3.8.1
mne==1.5.1
multidict==6.0.4
murmurhash==1.0.10
nbformat==5.9.2
neurokit2==0.2.7
pandas==2.0.3
pathtools==0.1.2
pillow==9.3.0
platformdirs==3.11.0
plotly==5.18.0
pooch==1.8.0
preshed==3.0.9
protobuf==4.25.0
psutil==5.9.6
py4j==0.10.9.7
pydantic==2.5.1
pydantic-core==2.14.3
pytorch-lightning==2.1.0
pyts==0.13.0
pywavelets==1.5.0
pyyaml==6.0.1
referencing==0.31.0
requests==2.28.1
rpds-py==0.13.1
scikit-base==0.6.1
scikit-learn==1.2.2
scipy==1.11.3
seaborn==0.13.0
sentry-sdk==1.34.0
setproctitle==1.3.3
shap==0.43.0
sktime==0.24.1
slicer==0.0.7
smart-open==6.4.0
smmap==5.0.1
soundfile==0.12.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
statsmodels==0.14.1
tenacity==8.2.3
thinc==8.2.1
torchmetrics==1.2.0
tsai==0.3.8
typer==0.9.0
tzdata==2023.3
urllib3==1.26.13
wandb==0.15.12
wasabi==1.1.2
weasel==0.3.4
wfdb==4.1.2
widgetsnbextension==4.0.9
wrapt==1.16.0
xgboost==2.0.2
yarl==1.9.2
